% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARQLResult.R
\docType{methods}
\name{dbSendQuery,SPARQLConnection-method}
\alias{dbSendQuery,SPARQLConnection-method}
\title{Send a query to SPARQL.}
\usage{
\S4method{dbSendQuery}{SPARQLConnection}(conn, statement, ...)
}
\description{
Send a query to SPARQL.
}

