% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARQLDriver.R, R/SPARQLConnection.R
\docType{methods}
\name{SPARQL}
\alias{SPARQL}
\alias{dbConnect,SPARQLDriver-method}
\title{Driver for SPARQL}
\usage{
SPARQL()

\S4method{dbConnect}{SPARQLDriver}(drv, endpoint, ...)
}
\arguments{
\item{drv}{An object created by \code{SPARQL()}}
}
\description{
This driver is for implementing the R database (DBI) API.

This driver is for implementing the R database (DBI) API.
}

