% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARQLDBI-package.R
\docType{package}
\name{SPARQLDBI}
\alias{SPARQLDBI}
\alias{SPARQLDBI-package}
\title{SPARQL As A Backend For DBI}
\description{
A simple implementation of R's DBI Interface. This provides access to SPARQL via DBI.
}

