% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARQLDriver.R
\docType{class}
\name{SPARQLDriver-class}
\alias{SPARQLDriver-class}
\alias{dbUnloadDriver,SPARQLDriver-method}
\title{SPARQL driver class}
\usage{
\S4method{dbUnloadDriver}{SPARQLDriver}(drv, ...)
}
\description{
SPARQL driver class
}
\keyword{internal}

